Split the PipelineGraph component into smaller ones
